backend/requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.18.4
altair==5.5.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
# audioop-lts==0.2.1 <- removed: requires Python >=3.13, we use 3.11
anyio==4.9.0
APScheduler==3.11.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arq==0.27.0
asgiref==3.11.1
astroid==4.0.4
asyncio==3.4.3
asyncio-throttle==1.0.2
attrs==25.3.0
Authlib==1.6.1
backoff==2.2.1
bandit==1.8.6
bcrypt==4.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
billiard==4.2.4
bitsandbytes==0.46.1
black==25.1.0
blinker==1.9.0
boolean.py==5.0
boto3==1.42.64
botocore==1.42.64
Brotli==1.1.0
build==1.3.0
CacheControl==0.14.4
cachetools==5.5.2
celery==5.6.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.3
coverage==7.13.4
cross-web==0.4.1
cryptography==45.0.5
cycler==0.12.1
cyclonedx-python-lib==8.9.0
dataclasses==0.6
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
detect-secrets==1.5.0
dill==0.4.1
discord.py==2.5.2
distro==1.9.0
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
dparse==0.6.4
durationpy==0.10
ecdsa==0.19.1
email-validator==2.3.0
execnet==2.1.2
fabric==3.2.2
factory_boy==3.3.3
Faker==33.1.0
fakeredis==2.33.0
fastapi==0.135.3
feedparser==6.0.11
ffmpy==0.6.1
filelock==3.16.1
flake8==7.3.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
flatbuffers==25.2.10
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
googleapis-common-protos==1.70.0
gradio==5.41.0
gradio_client==1.11.0
grafana-api==1.0.3
graphql-core==3.2.8
greenlet==3.3.2
groovy==0.1.2
grpcio==1.74.0
h11==0.16.0
hiredis==3.3.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.34.3
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
invoke==2.2.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.1.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
kombu==5.6.2
kubernetes==33.1.0
langchain-core==1.2.18
langchain-text-splitters==1.1.1
langgraph==1.1.0
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.10
langsmith==0.7.16
Levenshtein==0.27.3
license-expression==30.4.4
limits==5.8.0
lxml==6.0.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mccabe==0.7.0
mcp==1.26.0
mdurl==0.1.2
minio==7.2.20
mmh3==5.2.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.6.3
mypy==1.14.0
mypy_extensions==1.1.0
narwhals==2.0.1
networkx==3.5
nltk==3.9.1
nodeenv==1.10.0
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.98.0
opentelemetry-api==1.40.0
opentelemetry-distro==0.61b0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-instrumentation==0.61b0
opentelemetry-instrumentation-asgi==0.61b0
opentelemetry-instrumentation-fastapi==0.61b0
opentelemetry-instrumentation-httpx==0.61b0
opentelemetry-instrumentation-logging==0.61b0
opentelemetry-instrumentation-redis==0.61b0
opentelemetry-instrumentation-sqlalchemy==0.61b0
opentelemetry-instrumentation-system-metrics==0.61b0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opentelemetry-util-http==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packageurl-python==0.17.6
packaging==25.0
pandas==2.3.1
paramiko==3.5.1
passlib==1.7.4
pathlib2==2.3.7.post1
pathspec==0.12.1
pbr==6.1.1
pillow==11.3.0
ping3==5.1.5
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip-tools==7.5.3
pip_audit==2.8.0
platformdirs==4.3.8
plotly==6.2.0
pluggy==1.6.0
posthog==5.4.0
prisma==0.15.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.31.1
psutil==6.1.1
psycopg2-binary==2.9.10
py-cord==2.6.1
py-serializable==1.1.2
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycodestyle==2.14.0
pycparser==2.22
pycryptodome==3.23.0
pydantic>=2.10.0,<2.12
pydantic-settings>=2.13.1,<3.0
pydeck==0.9.1
pydub==0.25.1
pyflakes==3.4.0
pygame==2.6.1
PyGithub==2.9.0
Pygments==2.19.2
PyJWT==2.11.0
pylint==4.0.5
PyNaCl==1.5.0
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pytest==9.0.2
pytest-asyncio>=0.25.0
pytest-cov==7.1.0
pytest-xdist==3.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-jose==3.5.0
python-Levenshtein==0.27.3
python-memcached==1.62
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
# pywin32==311 <- REMOVED: Windows-only, not available in Linux Docker containers
PyYAML==6.0.2
qiskit==2.3.1
RapidFuzz==3.14.3
redis==5.3.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
# ruff bumped to 0.9.3 — gradio==5.41.0 requires ruff>=0.9.3
ruff==0.9.3
rustworkx==0.17.1
s3transfer==0.16.0
safehttpx==0.1.6
safetensors==0.5.3
# safety + safety-schemas removed from prod image — belongs in CI only (conflicts with mcp>=2.11 pydantic constraint)
schedule==1.2.2
scikit-learn==1.7.1
scipy==1.16.1
semantic-version==2.10.0
sentence-transformers==5.2.3
sentry-sdk==1.40.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
slowapi==0.1.9
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SQLAlchemy==2.0.48
sse-starlette==1.6.1
starlette==0.47.2
stevedore==5.4.1
strawberry-graphql==0.311.3
streamlit==1.48.0
structlog==24.1.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.55.2
typer==0.16.0
types-pyasn1==0.6.0.20250914
types-python-jose==3.5.0.20250531
types-requests==2.32.4.20260107
typing==3.7.4.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.5.0
uuid_utils==0.14.1
uvicorn==0.35.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.5.3
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.46.3
wrapt==1.17.2
wsproto==1.2.0
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
